+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Make cart subtotal data type from `integer` to `float`
+- Updated graphql schema version
+
 ## [1.2.3] - 2020-12-29
 
 ### Fixed